add service channel emulation support
The Bayeux protocol describe service channels (in the /service/ namespace)
for request/response style communication
(http://svn.cometd.org/trunk/bayeux/bayeux.html#toc_22).

The chat window uses two service channels (/service/chat/publish and
/service/chat/notify) to send and receive chat transcripts and system messages.

Faye don't support service channels (https://github.com/faye/faye/issues/172).

As a workaround, we emulate service channels using normal channels with a
unique name and only one subscriber each, of the form /service/.../uid
(uid is likely to be the client id). When the client connects (on
handshake), we subscribe the client to such a channel, that we can then
use to send messages back.

Note that the client expects the message coming through the service
channel, not the uniquely named one, so we strip the uid from the
channel name just before sending the message, after all the server side
routing has been resolved (see the diff of this commit).

One last thing: the Bayeux protocol says service channels are
unsubscribable, and so Faye will check the namespace of the channel and
prevent it. Since we think it's better to keep the channel name transformation
to a minimum, we still use the service namespace in the uniquely named channel
names, but we subscribe the client skipping validations with
(faye-adapter)._server._engine.subscribe(clientId, channel).

THIS CHANGE IS TO THE JAVASCRIPT IMPLEMENTATION ONLY.